+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The public API of this library consists of the functions declared in file
 
 ## [Unreleased]
 ### Fixed
-- `compact` handles zero length input correctly.
+- `compact` handles zero length input correctly. (#278)
 
 ## [3.6.0] - 2019-08-12
 ### Added