+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The public API of this library consists of the functions declared in file
 ### Fixed
 - `compact` handles zero length input correctly. (#278)
 - `bboxHexRadius` scaling factor adjusted to guarantee containment for `polyfill`. (#279)
+- `polyfill` new algorithm for up to 3x perf boost. (#282)
 
 ## [3.6.0] - 2019-08-12
 ### Added